+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - (Compendium) Acid Arrow was incorrectly set up to have a Ranged Weapon Attack
 - Fixed error messages for players when GM updated an entity which would do something with an actor's hit points they didn't own
 - NPC Lite sheets were bugging out majorly
+- Wounds and Vigor incorrectly checked for PC rules even for NPCs
 
 ### Changelog